+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## [0.4.0] - 2021-02-10
+### Added
+- Tab character can be used for indentation
+
+## [0.3.0]
 ## Added
 - Null values are not serialized
 - Program prints version
