+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "apish"
-version = "0.3.0"
+version = "0.4.0"
 authors = ["Fido <fidencio.garrido@gmail.com>"]
 edition = "2018"
